+++ NEWS.md
@@ -5,6 +5,7 @@ Bugfixes:
 
 - Explicitly settings DuckDB's `temp_directory` to the Andromeda temp folder to avoid running out of space in undefined temp location.
 
+- Set default memory limit to 20% instead of 80% of system memory to avoid out-of-memory errors when there are multiple Andromeda objects in memory.
 
 Andromeda 1.1.1
 ===============